@@ -11,7 +11,7 @@
   "license": "ISC",
   "homepage": "",
   "devDependencies": {
-    "@babel/core": "^7.26.7",
+    "@babel/core": "^7.28.4",
     "@babel/preset-env": "^7.26.7",
     "@babel/preset-react": "^7.26.3",
     "babel-loader": "^9.2.1",
@@ -24,8 +24,8 @@
   },
   "dependencies": {
     "prop-types": "^15.8.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "styled-components": "^5.1.0"
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
+    "styled-components": "^6.1.19"
   }
 }
